+++ projects/ngx-forms/src/lib/ngx-forms-ui/components/form-flow-wrapper/form-flow-wrapper.component.ts
@@ -1,13 +1,18 @@
-import {ChangeDetectionStrategy, Component, ViewEncapsulation} from '@angular/core';
+import {ChangeDetectionStrategy, Component, Input, ViewEncapsulation} from '@angular/core';
 
 @Component({
   selector: 'form-flow-wrapper',
   templateUrl: './form-flow-wrapper.component.html',
   styleUrls: ['./form-flow-wrapper.component.scss'],
   encapsulation: ViewEncapsulation.None,
-  changeDetection: ChangeDetectionStrategy.OnPush
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  host: {
+    '[style.grid-template-columns]': `'repeat(auto-fit, minmax(' + minWidth + 'px, 1fr))'`
+  }
 })
 export class FormFlowWrapperComponent {
+
+  @Input() minWidth = 200;
 
   constructor() { }