+++ projects/ngx-tools/src/lib/future-builder/directives/when-empty.directive.ts
@@ -1,4 +1,4 @@
-import {Directive, Input, OnDestroy, TemplateRef, ViewContainerRef} from '@angular/core';
+import {ChangeDetectorRef, Directive, Input, OnDestroy, TemplateRef, ViewContainerRef} from '@angular/core';
 import {map, switchMap} from "rxjs/operators";
 import {Subject, Subscription} from "rxjs";
 import {FutureError, FutureLoading} from "@consensus-labs/rxjs-tools";
@@ -18,7 +18,8 @@ export class WhenEmptyDirective<T> implements OnDestroy {
 
   constructor(
     private templateRef: TemplateRef<TemplateContext<T>>,
-    private viewContainer: ViewContainerRef
+    private viewContainer: ViewContainerRef,
+    private changes: ChangeDetectorRef
   ) {
     this.sub = this.states$.pipe(
       switchMap(x => x.empty$),
@@ -29,6 +30,7 @@ export class WhenEmptyDirective<T> implements OnDestroy {
     ).subscribe(c => {
       if (c) this.viewContainer.createEmbeddedView(this.templateRef, c);
       else this.viewContainer.clear();
+      this.changes.detectChanges();
     });
   }