+++ projects/ngx-tools/src/lib/future-builder/directives/when-loading-overlay.directive.ts
@@ -1,6 +1,6 @@
-import {Directive, Input, OnDestroy, TemplateRef, ViewContainerRef} from '@angular/core';
+import {ChangeDetectorRef, Directive, Input, OnDestroy, TemplateRef, ViewContainerRef} from '@angular/core';
 import {Subject, Subscription, switchMap} from "rxjs";
-import {distinctUntilChanged, map} from "rxjs/operators";
+import {map} from "rxjs/operators";
 import {FutureLoading} from "@consensus-labs/rxjs-tools";
 import {FutureSwitch} from "../models/future-switch.model";
 
@@ -16,7 +16,8 @@ export class WhenLoadingOverlayDirective<T> implements OnDestroy {
 
   constructor(
     private templateRef: TemplateRef<TemplateContext<T>>,
-    private viewContainer: ViewContainerRef
+    private viewContainer: ViewContainerRef,
+    private changes: ChangeDetectorRef
   ) {
     this.sub = this.states$.pipe(
       switchMap(x => x.loadingOverlay$),
@@ -27,6 +28,7 @@ export class WhenLoadingOverlayDirective<T> implements OnDestroy {
     ).subscribe(c => {
       if (c) this.viewContainer.createEmbeddedView(this.templateRef, c);
       else this.viewContainer.clear();
+      this.changes.detectChanges();
     });
   }