+++ projects/ngx-tools/src/lib/future-builder/directives/when-loading.directive.ts
@@ -1,6 +1,6 @@
-import {Directive, Input, OnDestroy, TemplateRef, ViewContainerRef} from '@angular/core';
+import {ChangeDetectorRef, Directive, Input, OnDestroy, TemplateRef, ViewContainerRef} from '@angular/core';
 import {Subject, Subscription, switchMap} from "rxjs";
-import {distinctUntilChanged, map} from "rxjs/operators";
+import {map} from "rxjs/operators";
 import {FutureLoading} from "@consensus-labs/rxjs-tools";
 import {FutureSwitch} from "../models/future-switch.model";
 
@@ -18,7 +18,8 @@ export class WhenLoadingDirective<T> implements OnDestroy {
 
   constructor(
     private templateRef: TemplateRef<TemplateContext<T>>,
-    private viewContainer: ViewContainerRef
+    private viewContainer: ViewContainerRef,
+    private changes: ChangeDetectorRef
   ) {
     this.sub = this.states$.pipe(
       switchMap(x => x.loading$),
@@ -29,6 +30,7 @@ export class WhenLoadingDirective<T> implements OnDestroy {
     ).subscribe(c => {
       if (c) this.viewContainer.createEmbeddedView(this.templateRef, c);
       else this.viewContainer.clear();
+      this.changes.detectChanges();
     });
   }